+++ example/lib/app_routes.dart
@@ -1,20 +1,24 @@
+import 'package:example/case_multiple_tab_screen/tab_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_link_nav/flutter_link_nav.dart';
 
-import 'detail_screen.dart';
-import 'main.dart';
-import 'main_screen.dart';
+import 'case_normal/detail_screen.dart';
+import 'case_normal/main.dart';
+import 'case_normal/main_screen.dart';
 
 class ExampleAppRoutes extends AppRoutes {
   static const String mainScreen = MainScreen.routeName;
   static const String detailScreen = DetailScreen.routeName;
+
+  static const String tabScreen = TabScreen.routeName;
+  static const String another = AnotherScreen.routeName;
 
   @override
   Map<String, RouteConfig> get routes => {
     mainScreen: RouteConfig(
       widgetRegister: (queryParams) => const MainScreen(),
     ),
-    detailScreen: RouteConfig(
+    '$mainScreen/$detailScreen': RouteConfig(
       widgetRegister: (queryParams) => const DetailScreen(),
     ),
     'sheet': RouteConfig(
@@ -34,5 +38,15 @@ class ExampleAppRoutes extends AppRoutes {
         );
       },
     ),
+    tabScreen: RouteConfig(
+      widgetRegister: (queryParams) =>
+          TabScreen(route: queryParams != null ? queryParams['tab'] : null),
+    ),
+    another: RouteConfig(
+      widgetRegister: (queryParams) => const AnotherScreen(),
+    ),
   };
+
+  @override
+  Set<String>? get tabBasedRoutes => {tabScreen};
 }